convert stake script to typescript and extend to get validators
